refactor: simplify ProtectedRoute and implement AuthProvider for authentication handling
